Move UriMatcher related annotations to their own package
